refactor: move footnote refs in data cells to own component
